Resolve signer display from settings and separate wallet switch ownership
