+++ app/ts/components/activeAddressChange.ts
@@ -6,9 +6,8 @@ export async function requestActiveAddressChange(
 	activeAddress: bigint | 'signer',
 	simulationMode: boolean,
 	sendMessage: (message: ChangeActiveAddress) => Promise<ChangeActiveAddressReply | undefined> = sendPopupMessageWithReply,
-	addressChangeRequestId?: string,
 ) {
-	const reply = await sendMessage({ method: 'popup_changeActiveAddress', data: { activeAddress, simulationMode, ...(addressChangeRequestId === undefined ? {} : { addressChangeRequestId }) } })
+	const reply = await sendMessage({ method: 'popup_changeActiveAddress', data: { activeAddress, simulationMode } })
 	if (reply === undefined) throw new Error(getMissingPopupReplyErrorMessage('Changing the active address'))
 	if (!reply.ok) throw new Error(reply.message)
 }
